+++ sample/src/main/java/pl/charmas/android/reactivelocation/sample/PlacesActivity.java
@@ -2,8 +2,9 @@ package pl.charmas.android.reactivelocation.sample;
 
 import android.location.Location;
 import android.os.Bundle;
-import android.support.v7.app.ActionBarActivity;
 import android.text.TextUtils;
+import android.view.View;
+import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.ListView;
@@ -15,22 +16,22 @@ import com.google.android.gms.location.places.PlaceLikelihood;
 import com.google.android.gms.location.places.PlaceLikelihoodBuffer;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
+import com.jakewharton.rxbinding.widget.RxTextView;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import pl.charmas.android.reactivelocation.DataBufferObservable;
 import pl.charmas.android.reactivelocation.ReactiveLocationProvider;
-import pl.charmas.android.reactivelocation.sample.utils.TextObservable;
 import rx.Observable;
 import rx.functions.Action1;
 import rx.functions.Func1;
 import rx.functions.Func2;
 import rx.subscriptions.CompositeSubscription;
 
-import static rx.android.app.AppObservable.bindActivity;
+import static pl.charmas.android.reactivelocation.sample.utils.UnsubscribeIfPresent.unsubscribe;
 
-public class PlacesActivity extends ActionBarActivity {
+public class PlacesActivity extends BaseActivity {
 
     private TextView currentPlaceView;
     private EditText queryView;
@@ -45,34 +46,42 @@ public class PlacesActivity extends ActionBarActivity {
         currentPlaceView = (TextView) findViewById(R.id.current_place_view);
         queryView = (EditText) findViewById(R.id.place_query_view);
         placeSuggestionsList = (ListView) findViewById(R.id.place_suggestions_list);
+        placeSuggestionsList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                AutocompleteInfo info = (AutocompleteInfo) parent.getAdapter().getItem(position);
+                startActivity(PlacesResultActivity.getStartIntent(PlacesActivity.this, info.id));
+            }
+        });
 
         reactiveLocationProvider = new ReactiveLocationProvider(this);
     }
 
     @Override
-    protected void onStart() {
-        super.onStart();
+    protected void onLocationPermissionGranted() {
         compositeSubscription = new CompositeSubscription();
         compositeSubscription.add(
-                bindActivity(this, reactiveLocationProvider.getCurrentPlace(null))
-                        .flatMap(new Func1<PlaceLikelihoodBuffer, Observable<PlaceLikelihood>>() {
+                reactiveLocationProvider.getCurrentPlace(null)
+                        .subscribe(new Action1<PlaceLikelihoodBuffer>() {
                             @Override
-                            public Observable<PlaceLikelihood> call(PlaceLikelihoodBuffer placeLikelihoods) {
-                                return DataBufferObservable.from(placeLikelihoods);
-                            }
-                        })
-                        .firstOrDefault(null)
-                        .subscribe(new Action1<PlaceLikelihood>() {
-                            @Override
-                            public void call(PlaceLikelihood place) {
-                                if (place != null) {
-                                    currentPlaceView.setText(place.getPlace().getName());
+                            public void call(PlaceLikelihoodBuffer buffer) {
+                                PlaceLikelihood likelihood = buffer.get(0);
+                                if (likelihood != null) {
+                                    currentPlaceView.setText(likelihood.getPlace().getName());
                                 }
+                                buffer.release();
                             }
                         })
         );
 
-        Observable<String> queryObservable = TextObservable.create(queryView)
+        Observable<String> queryObservable = RxTextView
+                .textChanges(queryView)
+                .map(new Func1<CharSequence, String>() {
+                    @Override
+                    public String call(CharSequence charSequence) {
+                        return charSequence.toString();
+                    }
+                })
                 .debounce(1, TimeUnit.SECONDS)
                 .filter(new Func1<String, Boolean>() {
                     @Override
@@ -81,7 +90,7 @@ public class PlacesActivity extends ActionBarActivity {
                     }
                 });
         Observable<Location> lastKnownLocationObservable = reactiveLocationProvider.getLastKnownLocation();
-        Observable<List<String>> suggestionsObservable = Observable
+        Observable<AutocompletePredictionBuffer> suggestionsObservable = Observable
                 .combineLatest(queryObservable, lastKnownLocationObservable, new Func2<String, Location, QueryWithCurrentLocation>() {
                     @Override
                     public QueryWithCurrentLocation call(String query, Location currentLocation) {
@@ -100,22 +109,17 @@ public class PlacesActivity extends ActionBarActivity {
                         );
                         return reactiveLocationProvider.getPlaceAutocompletePredictions(q.query, bounds, null);
                     }
-                }).flatMap(new Func1<AutocompletePredictionBuffer, Observable<List<String>>>() {
-                    @Override
-                    public Observable<List<String>> call(AutocompletePredictionBuffer autocompletePredictions) {
-                        return DataBufferObservable.from(autocompletePredictions).map(new Func1<AutocompletePrediction, String>() {
-                            @Override
-                            public String call(AutocompletePrediction autocompletePrediction) {
-                                return autocompletePrediction.getDescription();
-                            }
-                        }).toList();
-                    }
                 });
 
-        compositeSubscription.add(bindActivity(this, suggestionsObservable).subscribe(new Action1<List<String>>() {
+        compositeSubscription.add(suggestionsObservable.subscribe(new Action1<AutocompletePredictionBuffer>() {
             @Override
-            public void call(List<String> autocompletePredictions) {
-                placeSuggestionsList.setAdapter(new ArrayAdapter<>(PlacesActivity.this, android.R.layout.simple_list_item_1, autocompletePredictions));
+            public void call(AutocompletePredictionBuffer buffer) {
+                List<AutocompleteInfo> infos = new ArrayList<>();
+                for (AutocompletePrediction prediction : buffer) {
+                    infos.add(new AutocompleteInfo(prediction.getFullText(null).toString(), prediction.getPlaceId()));
+                }
+                buffer.release();
+                placeSuggestionsList.setAdapter(new ArrayAdapter<>(PlacesActivity.this, android.R.layout.simple_list_item_1, infos));
             }
         }));
     }
@@ -123,8 +127,7 @@ public class PlacesActivity extends ActionBarActivity {
     @Override
     protected void onStop() {
         super.onStop();
-        compositeSubscription.unsubscribe();
-        compositeSubscription = null;
+        unsubscribe(compositeSubscription);
     }
 
     private static class QueryWithCurrentLocation {
@@ -134,6 +137,21 @@ public class PlacesActivity extends ActionBarActivity {
         private QueryWithCurrentLocation(String query, Location location) {
             this.query = query;
             this.location = location;
+        }
+    }
+
+    private static class AutocompleteInfo {
+        private final String description;
+        private final String id;
+
+        private AutocompleteInfo(String description, String id) {
+            this.description = description;
+            this.id = id;
+        }
+
+        @Override
+        public String toString() {
+            return description;
         }
     }
 }